**made dynamo db initialization optional
