+++ docker-compose.yml
@@ -17,8 +17,6 @@ services:
       - BOT_SERVICE__TIMEOUTSECONDS=5
       - LOG_LEVEL__DEFAULT=Debug
       - LOG_LEVEL__MICROSOFTASPNETCORE=Debug
-      - AWS__REGION=eu-central-1
-      - AWS__DYNAMODB_TABLENAME=GammonX
     depends_on:
       - bot-service
