+++ src/GammonX/GammonX.Server/.env
@@ -12,8 +12,8 @@ BOT_SERVICE__TIMEOUTSECONDS=5
 GAME_SERVICE__BASEPATH=__undefined__
 
 # AWS Settings
-AWS__AWS_ACCESS_KEY_ID=__undefined__
-AWS__AWS_SECRET_ACCESS_KEY=__undefined__
-AWS__REGION=__undefined__
-AWS__DYNAMODB_SERVICEURL=__undefined__
-AWS__DYNAMODB_TABLENAME=__undefined__
+# AWS__AWS_ACCESS_KEY_ID=__undefined__
+# AWS__AWS_SECRET_ACCESS_KEY=__undefined__
+# AWS__REGION=__undefined__
+# AWS__DYNAMODB_SERVICEURL=__undefined__
+# AWS__DYNAMODB_TABLENAME=__undefined__
