+++ src/GammonX/GammonX.Server/.env.local
@@ -12,8 +12,8 @@ BOT_SERVICE__TIMEOUTSECONDS=5
 GAME_SERVICE__BASEPATH=/game
 
 # AWS Settings
-AWS__AWS_ACCESS_KEY_ID=local
-AWS__AWS_SECRET_ACCESS_KEY=local
-AWS__REGION=
-AWS__DYNAMODB_SERVICEURL=http://localhost:8000
-AWS__DYNAMODB_TABLENAME=GammonX
+# AWS__AWS_ACCESS_KEY_ID=local
+# AWS__AWS_SECRET_ACCESS_KEY=local
+# AWS__REGION=
+# AWS__DYNAMODB_SERVICEURL=http://localhost:8000
+# AWS__DYNAMODB_TABLENAME=GammonX
