--- src/GammonX/GammonX.Server/Controllers/PlayersController.cs
@@ -1,17 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace GammonX.Server.Controllers
-{
-	[ApiController]
-	[Route("api/[controller]")]
-	public class PlayersController : Controller
-	{
-		[HttpGet]
-		public IActionResult Get()
-		{
-			return Ok();
-		}
-
-
-	}
-}
