+++ src/GammonX/GammonX.Server/Program.cs
@@ -1,7 +1,4 @@
-using Amazon;
 using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.DataModel;
-using Amazon.Runtime;
 using DotNetEnv;
 
 using GammonX.Engine.Services;
@@ -13,7 +10,7 @@ using GammonX.Server.Data.DynamoDb;
 using GammonX.Server.Data.Repository;
 using GammonX.Server.Models;
 using GammonX.Server.Services;
-
+using GammonX.Server.Services.extensions;
 using Microsoft.Extensions.Options;
 
 using Serilog;
@@ -51,48 +48,8 @@ builder.Services.Configure<GameServiceOptions>(
 	builder.Configuration.GetSection("GAME_SERVICE"));
 // -------------------------------------------------------------------------------
 // DATABASE SETUP
-// -------------------------------------------------------------------------------
-builder.Services.Configure<AwsServiceOptions>(
-	builder.Configuration.GetSection("AWS"));
-
-builder.Services.AddSingleton<IAmazonDynamoDB>(sp =>
-{
-	var options = sp.GetRequiredService<IOptions<AwsServiceOptions>>().Value;
-	var isLocal = string.IsNullOrEmpty(options.REGION);
-	if (isLocal)
-	{
-		// local docker instance
-		var accessKeyId = options.AWS_ACCESS_KEY_ID;
-		var secretAccessKey = options.AWS_SECRET_ACCESS_KEY;
-		var config = new AmazonDynamoDBConfig
-		{
-			ServiceURL = options.DYNAMODB_SERVICEURL
-		};
-		var credentials = new BasicAWSCredentials(options.AWS_ACCESS_KEY_ID, options.AWS_SECRET_ACCESS_KEY);
-		return new AmazonDynamoDBClient(credentials, config);
-	}
-	else
-	{
-		// AWS hosted
-		var region = RegionEndpoint.GetBySystemName(options.REGION);
-		var config = new AmazonDynamoDBConfig
-		{
-			// We do not need a specific service url, the region endpoint is sufficient
-			RegionEndpoint = region
-		};
-		// We use role based auth
-		return new AmazonDynamoDBClient(config);
-	}
-});
-builder.Services.AddSingleton<IDynamoDBContext>(sp =>
-{
-	var client = sp.GetRequiredService<IAmazonDynamoDB>();
-	var contextBuilder = new DynamoDBContextBuilder();
-	contextBuilder.WithDynamoDBClient(() => client);
-	return contextBuilder.Build();
-});
-
-builder.Services.AddScoped<IPlayerRepository, DynamoDbPlayerRepository>();
+var awsConfig = builder.Configuration.GetSection("AWS");
+builder.Services.AddConditionalDynamoDb(awsConfig);
 // -------------------------------------------------------------------------------
 // DEPENDENCY INJECTION
 // -------------------------------------------------------------------------------
@@ -176,8 +133,11 @@ Log.Information("AWS DYNAMO DB SERVICE URL: {DynamoDbServiceUrl}", Environment.G
 using (var scope = app.Services.CreateScope())
 {
 	var options = app.Services.GetRequiredService<IOptions<AwsServiceOptions>>().Value;
-	var dynamoClient = scope.ServiceProvider.GetRequiredService<IAmazonDynamoDB>();
-	await DynamoDbInitializer.EnsureTablesExistAsync(dynamoClient, options);
+	if (options.Required)
+	{
+		var dynamoClient = scope.ServiceProvider.GetRequiredService<IAmazonDynamoDB>();
+		await DynamoDbInitializer.EnsureTablesExistAsync(dynamoClient, options);
+	}
 }
 
 app.Run();
