+++ src/GammonX/GammonX.Server/Services/extensions/DynamoDbServiceExtensions.cs
@@ -1,0 +1,65 @@
+﻿using Amazon;
+using Amazon.DynamoDBv2;
+using Amazon.DynamoDBv2.DataModel;
+using Amazon.Runtime;
+
+using GammonX.Server.Data.DynamoDb;
+using GammonX.Server.Data.Repository;
+
+using Microsoft.Extensions.Options;
+
+namespace GammonX.Server.Services.extensions
+{
+	public static class DynamoDbServiceExtensions
+	{
+		public static void AddConditionalDynamoDb(this IServiceCollection services, IConfigurationSection config)
+		{
+			services.Configure<AwsServiceOptions>(config);
+			// we check manually if a dynamo config is required
+			var tableName = Environment.GetEnvironmentVariable("AWS__DYNAMODB_TABLENAME");
+			if (string.IsNullOrEmpty(tableName))
+			{
+				return;
+			}
+
+			services.AddSingleton<IAmazonDynamoDB>(sp =>
+			{
+				var options = sp.GetRequiredService<IOptions<AwsServiceOptions>>().Value;
+				var isLocal = string.IsNullOrEmpty(options.REGION);
+				if (isLocal)
+				{
+					// local docker instance
+					var accessKeyId = options.AWS_ACCESS_KEY_ID;
+					var secretAccessKey = options.AWS_SECRET_ACCESS_KEY;
+					var config = new AmazonDynamoDBConfig
+					{
+						ServiceURL = options.DYNAMODB_SERVICEURL
+					};
+					var credentials = new BasicAWSCredentials(options.AWS_ACCESS_KEY_ID, options.AWS_SECRET_ACCESS_KEY);
+					return new AmazonDynamoDBClient(credentials, config);
+				}
+				else
+				{
+					// AWS hosted
+					var region = RegionEndpoint.GetBySystemName(options.REGION);
+					var config = new AmazonDynamoDBConfig
+					{
+						// We do not need a specific service url, the region endpoint is sufficient
+						RegionEndpoint = region
+					};
+					// We use role based auth
+					return new AmazonDynamoDBClient(config);
+				}
+			});
+			services.AddSingleton<IDynamoDBContext>(sp =>
+			{
+				var client = sp.GetRequiredService<IAmazonDynamoDB>();
+				var contextBuilder = new DynamoDBContextBuilder();
+				contextBuilder.WithDynamoDBClient(() => client);
+				return contextBuilder.Build();
+			});
+
+			services.AddScoped<IPlayerRepository, DynamoDbPlayerRepository>();
+		}
+	}
+}
